Module Flextesa.Running_processes

Run and monitor processes.

module Process : sig ... end

The definition of a process, for now, a process within a process-group or a Docker container.

module State : sig ... end

The container for a list of running or not processes.

val output_path : < paths : Paths.t; .. > -> Process.t -> [ `Meta | `Stderr | `Stdout ] -> Internal_pervasives.string

Return the path (within Paths's root-path) where the process writes its output or metadata.

val ef_processes : < paths : Paths.t; .. > -> State.process_state Internal_pervasives.list -> Easy_format.t

val ef : ?⁠all:Internal_pervasives.bool -> < runner : State.t; .. > -> Easy_format.t

val start : < application_name : Internal_pervasives.string; paths : Paths.t; runner : State.t; .. > -> Process.t -> (State.process_state, [> Internal_pervasives.System_error.t ]) Internal_pervasives.Asynchronous_result.t

val wait : < runner : State.t; .. > -> State.process_state -> (Lwt_unix.process_status, [> Internal_pervasives.System_error.t ]) Internal_pervasives.Asynchronous_result.t

val kill : < runner : State.t; .. > -> State.process_state -> (Internal_pervasives.unit, [> Internal_pervasives.System_error.t ]) Internal_pervasives.Asynchronous_result.t

val wait_all : < runner : State.t; .. > -> (Internal_pervasives.unit, [> Internal_pervasives.System_error.t ]) Internal_pervasives.Asynchronous_result.t

val kill_all : < runner : State.t; .. > -> (Internal_pervasives.unit, [> Internal_pervasives.System_error.t ]) Internal_pervasives.Asynchronous_result.t

val find_process_by_id : ?⁠only_running:Internal_pervasives.bool -> < runner : State.t; .. > -> f:(Internal_pervasives.string -> Internal_pervasives.bool) -> (State.process_state Internal_pervasives.list, [> ]) Internal_pervasives.Asynchronous_result.t

val run_cmdf : ?⁠id_prefix:Internal_pervasives.string -> < paths : Paths.t; runner : State.t; .. > Internal_pervasives.Base_state.t -> ('a, Internal_pervasives.unit, Internal_pervasives.string, (Internal_pervasives.Process_result.t, [> Internal_pervasives.System_error.t ]) Internal_pervasives.Asynchronous_result.t) Internal_pervasives.format4 -> 'a

Run a shell command and wait for its end.

val run_successful_cmdf : < paths : Paths.t; runner : State.t; .. > Internal_pervasives.Base_state.t -> ('a, Internal_pervasives.unit, Internal_pervasives.string, (Internal_pervasives.Process_result.t, [> Internal_pervasives.System_error.t | Internal_pervasives.Process_result.Error.t ]) Internal_pervasives.Asynchronous_result.t) Internal_pervasives.format4 -> 'a

val run_genspio : < paths : Paths.t; runner : State.t; .. > Internal_pervasives.Base_state.t -> Internal_pervasives.string -> 'a Genspio.Language.t -> (Lwt_unix.process_status, [> Internal_pervasives.System_error.t ]) Internal_pervasives.Asynchronous_result.t

module Async : sig ... end