Module Running_processes.Async

val run_cmdf : ?⁠id_base:Internal_pervasives.string -> < runner : State.t; .. > -> f:(State.process_state -> Lwt_process.process_full -> ('return_ok, [> Internal_pervasives.System_error.t ] as 'error) Internal_pervasives.Asynchronous_result.t) -> ('c, Internal_pervasives.unit, Internal_pervasives.string, (Unix.process_status * 'return_ok, 'error) Internal_pervasives.Asynchronous_result.t) Internal_pervasives.format4 -> 'c

Run a shell command and run a function over the process data before waiting for its end.

val fold_process : Lwt_process.process_full -> init:'a -> f:('a -> Internal_pervasives.string -> Internal_pervasives.string -> ([< `Continue of 'a | `Done of 'a ], [> Internal_pervasives.System_error.t ] as 'b) Internal_pervasives.Asynchronous_result.t) -> ('a, 'b) Internal_pervasives.Asynchronous_result.t

Fold over the output and error-output of a Lwt_process.process_full, and