Module Test_scenario.Network

Start networks from (and manipulate) Topology.t values.

type t = private {

nodes : Tezos_node.t Internal_pervasives.list;

}

val make : Tezos_node.t Internal_pervasives.list -> t

val start_up : ?⁠do_activation:Internal_pervasives.bool -> ?⁠check_ports:Internal_pervasives.bool -> < Internal_pervasives.Base_state.base; env_config : Environment_configuration.t; paths : Paths.t; console : Console.t; runner : Running_processes.State.t; .. > -> client_exec:Tezos_executable.t -> t -> (Internal_pervasives.unit, [> `Empty_protocol_list | Internal_pervasives.System_error.t | Internal_pervasives.Process_result.Error.t | Internal_pervasives.Process_result.Error.t | `Too_many_protocols of Tezos_protocol.t Internal_pervasives.list | `Too_many_timestamp_delays of Tezos_protocol.t Internal_pervasives.list | `Waiting_for of Internal_pervasives.string * [ `Time_out ] ]) Internal_pervasives.Asynchronous_result.t

Start the nodes, bootstrap the client, and (potentially) activate the protocol of a network.

?do_activation: if true the start-up will try to activate the protocol, if false it will do it only if the current level of the chain is 0.