Module Test_scenario.Queries

Run queries on running networks.

val all_levels : ?⁠chain:Internal_pervasives.string -> < application_name : Internal_pervasives.string; console : Console.t; paths : Paths.t; runner : Running_processes.State.t; .. > -> nodes:Tezos_node.t Internal_pervasives.list -> ((Internal_pervasives.string * [> `Failed | `Level of Internal_pervasives.int | `Null | `Unknown of Internal_pervasives.string ]) Internal_pervasives.list, [> Internal_pervasives.System_error.t ]) Internal_pervasives.Asynchronous_result.t

Get the current chain level for all the nodes, returns node-ID × level values.

val wait_for_all_levels_to_be : ?⁠attempts_factor:Internal_pervasives.float -> ?⁠chain:Internal_pervasives.string -> < application_name : Internal_pervasives.string; console : Console.t; paths : Paths.t; runner : Running_processes.State.t; .. > -> attempts:Internal_pervasives.int -> seconds:Internal_pervasives.float -> Tezos_node.t Internal_pervasives.list -> [< `At_least of Internal_pervasives.int | `Equal_to of Internal_pervasives.int ] -> (Internal_pervasives.unit, [> Internal_pervasives.System_error.t | `Waiting_for of Internal_pervasives.string * [ `Time_out ] ]) Internal_pervasives.Asynchronous_result.t

Try-sleep-loop waiting for all given nodes to reach a given level.